2021/2/18

+++ SourceCode/kernel/initializeLDTDescriptorInGDT.c
@@ -13,7 +13,7 @@ PUBLIC void initializeLDT() {
 		u32 base = (u32)gdt[INDEX_FLAT_RW].baseLow;
 		base = base | (gdt[INDEX_FLAT_RW].basehigh1 << 16);
 		base = base | (gdt[INDEX_FLAT_RW].baseHigh2 << 24);
-		base += &(pointer->ldt);
+		base += (u32)&(pointer->ldt);
 		u32 limit = sizeof(pointer->ldt) - 1;
 		u16 attribute = DA_32 | DA_P | DA_NS | DA_LDT;
 		initializeDescriptor(target, base, limit, attribute);
