+++ SourceCode/kernel/initializeTSSDescriptorInGDT.c
@@ -12,7 +12,7 @@ PUBLIC void initializeTSS() {
 	u32 base = (u32)gdt[INDEX_FLAT_RW].baseLow;
 	base = base | (gdt[INDEX_FLAT_RW].basehigh1 << 16);
 	base = base | (gdt[INDEX_FLAT_RW].baseHigh2 << 24);
-	base += &tss;
+	base += (u32)&tss;
 	u32 limit = sizeof(tss) - 1;
 	u16 attribute = DA_32 | DA_P | DA_NS | DA_386TSS;
 	initializeDescriptor(target, base, limit, attribute);
